api/pyrate/pyproject.toml
[tool.poetry]
name = "pyrate"
version = "0.1.0"
description = ""
authors = ["Hevon Gordon <hevondgordon.professional@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
black = "^21.9b0"
Flask = "^2.0.2"
SQLAlchemy = "^1.4.26"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.1"
python-decouple = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
